[browser tests] click on theia app shell and clear local browser storage
Clicking on the Theia application shell before starting running the tests helps
with test failures when running the suite with a UI (i.e. non-headless mode).
For some reason unless we do that, something is not set correctly, and most tests
that rely querying the UI fail.

Also clearing the local browser storage, to possibly avoid starting the app
with some state from the previous run. This could help when running the tests
locally, since CI should in theory always start with a clean environment.

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>